migrate mooc-header organism to testing-library
